--- a/sdk/nodejs/datasync/efsLocation.ts
+++ b/sdk/nodejs/datasync/efsLocation.ts
@@ -6,7 +6,7 @@ import * as inputs from "../types/input";
 import * as outputs from "../types/output";
 import * as utilities from "../utilities";
 
-import {ARN} from "..";
+import {ARN} from "../arn";
 
 /**
  * Manages an AWS DataSync EFS Location.