--- a/sdk/nodejs/ec2transitgateway/transitGateway.ts
+++ b/sdk/nodejs/ec2transitgateway/transitGateway.ts
@@ -4,7 +4,7 @@
 import * as pulumi from "@pulumi/pulumi";
 import * as utilities from "../utilities";
 
-import {ARN} from "..";
+import {ARN} from "../arn";
 
 /**
  * Manages an EC2 Transit Gateway.