--- a/sdk/nodejs/iam/rolePolicyAttachment.ts
+++ b/sdk/nodejs/iam/rolePolicyAttachment.ts
@@ -4,7 +4,7 @@
 import * as pulumi from "@pulumi/pulumi";
 import * as utilities from "../utilities";
 
-import {ARN} from "..";
+import {ARN} from "../arn";
 import {Role} from "./index";
 
 /**