--- a/sdk/nodejs/iam/userPolicyAttachment.ts
+++ b/sdk/nodejs/iam/userPolicyAttachment.ts
@@ -4,7 +4,7 @@
 import * as pulumi from "@pulumi/pulumi";
 import * as utilities from "../utilities";
 
-import {ARN} from "..";
+import {ARN} from "../arn";
 import {User} from "./index";
 
 /**