--- a/sdk/nodejs/iot/thingPrincipalAttachment.ts
+++ b/sdk/nodejs/iot/thingPrincipalAttachment.ts
@@ -4,7 +4,7 @@
 import * as pulumi from "@pulumi/pulumi";
 import * as utilities from "../utilities";
 
-import {ARN} from "..";
+import {ARN} from "../arn";
 
 /**
  * Attaches Principal to AWS IoT Thing.