--- a/sdk/nodejs/sns/topic.ts
+++ b/sdk/nodejs/sns/topic.ts
@@ -4,7 +4,7 @@
 import * as pulumi from "@pulumi/pulumi";
 import * as utilities from "../utilities";
 
-import {ARN} from "..";
+import {ARN} from "../arn";
 
 /**
  * Provides an SNS topic resource