…set of TypeScript files

This increases build performance dramatically.

For a sample project which created a S3 Bucket and a Lambda here


##Before
```
import * as pulumi from "@pulumi/pulumi";
import * as aws from "@pulumi/aws";

export const blogsTable = new aws.dynamodb.Table('some-table', {
    attributes: [
        {
            name: 'PK',
            type: 'S',
        },
        {
            name: 'SK',
            type: 'S',
        },
    ],
    billingMode: 'PAY_PER_REQUEST',
    hashKey: 'PK',
    rangeKey: 'SK',
    readCapacity: 5,
    writeCapacity: 5,
})

// Create an AWS resource (S3 Bucket)
const bucket = new aws.s3.Bucket("my-bucket");

// Export the name of the bucket
export const bucketName = bucket.id;

new aws.lambda.CallbackFunction('callback', {
    callback: async () => {
    },
})
```

Stats

```
pulumi-module-example node_modules/.bin/tsc --extendedDiagnostics
Files:                       1466
Lines:                     616374
Nodes:                    1713503
Identifiers:               616003
Symbols:                   388012
Types:                     146551
Memory used:              675092K
Assignability cache size:   45298
Identity cache size:            3
Subtype cache size:             0
I/O Read time:              0.79s
Parse time:                 2.84s
Program time:               4.50s
Bind time:                  0.91s
Check time:                 3.16s
transformTime time:         0.01s
Source Map time:            0.00s
commentTime time:           0.00s
I/O Write time:             0.00s
printTime time:             0.01s
Emit time:                  0.01s
Total time:                 8.59
```

## After 

```
import * as pulumi from "@pulumi/pulumi";
import * as s3 from "@pulumi/aws/s3";
import * as lambda from '@pulumi/aws/lambda'
import * as dynamodb from '@pulumi/aws/dynamodb'

export const blogsTable = new dynamodb.Table('some-table', {
    attributes: [
        {
            name: 'PK',
            type: 'S',
        },
        {
            name: 'SK',
            type: 'S',
        },
    ],
    billingMode: 'PAY_PER_REQUEST',
    hashKey: 'PK',
    rangeKey: 'SK',
    readCapacity: 5,
    writeCapacity: 5,
})

// Create an AWS resource (S3 Bucket)
const bucket = new s3.Bucket("my-bucket");

// Export the name of the bucket
export const bucketName = bucket.id;

new lambda.CallbackFunction('callback', {
    callback: async () => {
    },
})
```

Stats
```
pulumi-module-example node_modules/.bin/tsc --extendedDiagnostics 
Files:                        210
Lines:                      87083
Nodes:                     275249
Identifiers:                96759
Symbols:                    76109
Types:                      25847
Memory used:              139730K
Assignability cache size:   38203
Identity cache size:            3
Subtype cache size:             0
I/O Read time:              0.02s
Parse time:                 0.47s
Program time:               0.64s
Bind time:                  0.26s
Check time:                 1.23s
transformTime time:         0.01s
Source Map time:            0.00s
commentTime time:           0.00s
I/O Write time:             0.00s
printTime time:             0.01s
Emit time:                  0.01s
Total time:                 2.15s
```

The AWS-SDK has similar issues, but is a bit harder to unpick