@stack72 2020-06-24
> @stack72 another issue i spotted is the package.json version is `v2.10.0`, I was going to use https://github.com/ds300/patch-package to fix this locally in my projects but it doesn't work on `@pulumi/aws` because of the `v` in version. It should just be `2.10.0`

mmmhhhh I have a feeling we have some scripts that strip that V before we publish it :/