@JakeGinnivan 2020-06-24
@stack72 dang, this approach is broken. https://github.com/pulumi/pulumi-aws/issues/772

Not going to be able to use this for now, unless #772 is also fixed