@JakeGinnivan 2020-06-29
@stack72 who could sort out #772, would make a decent difference to anyone using TypeScript project references during build